Implement better thread job process completion polling.
This will likely fix any segmentation fault issues (as long as they're not to do with silent thread crashes).

The main consensus reached in regards to the segmentation fault was a null pointer. The RecordingSaveTask checks the thread every tick. Between two ticks, there is a difference of an ideal 0.05s. It's possible that the thread completes its task and then destroys itself freeing up the memory that was previously being used by it. Then when the task tries to access it, it points to an unassigned memory which results in the segmentation fault.